*: move srv into pkg discovery
