+++ pkg/testutil/testutil.go
@@ -17,7 +17,11 @@
 package testutil
 
 import (
+	"net/url"
 	"runtime"
+	"testing"
+
+	"github.com/coreos/etcd/pkg/types"
 )
 
 // WARNING: This is a hack.
@@ -27,4 +31,15 @@ func ForceGosched() {
 	for i := 0; i < 10000; i++ {
 		runtime.Gosched()
 	}
+}
+
+func MustNewURLs(t *testing.T, urls []string) []url.URL {
+	if urls == nil {
+		return nil
+	}
+	u, err := types.NewURLs(urls)
+	if err != nil {
+		t.Fatalf("unexpected new urls error: %v", err)
+	}
+	return u
 }
